+++ starkinfra/creditnote/__transfer.py
@@ -4,7 +4,7 @@ from starkcore.utils.checks import check_datetime, check_datetime_or_date
 
 class Transfer(Resource):
     """# creditnote.Transfer object
-    Transfer object to be created after contract signature and sent to the credit receiver.
+    Transfer sent to the credit receiver after the contract is signed.
     ## Parameters (required):
     - name [string]: receiver full name. ex: "Anthony Edward Stark"
     - tax_id [string]: receiver tax ID (CPF or CNPJ) with or without formatting. ex: "01234567890" or "20.018.183/0001-80"
@@ -16,15 +16,15 @@ class Transfer(Resource):
     - tags [list of strings]: list of strings for reference when searching for transfers. ex: ["employees", "monthly"]
     ## Attributes (return-only):
     - amount [integer]: amount in cents to be transferred. ex: 1234 (= R$ 12.34)
-    - external_id [string, default None]: url safe string that must be unique among all your transfers. Duplicated external_ids will cause failures. By default, this parameter will block any transfer that repeats amount and receiver information on the same date. ex: "my-internal-id-123456"
-    - scheduled [datetime.date, datetime.datetime or string, default now]: date or datetime when the transfer will be processed. May be pushed to next business day if necessary. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
-    - description [string, default None]: optional description to override default description to be shown in the bank statement. ex: "Payment for service #1234"
-    - id [string, default None]: unique id returned when the transfer is created. ex: "5656565656565656"
-    - fee [integer, default None]: fee charged when the Transfer is processed. ex: 200 (= R$ 2.00)
-    - status [string, default None]: current transfer status. ex: "success" or "failed"
-    - transaction_ids [list of strings, default None]: ledger Transaction IDs linked to this Transfer (if there are two, the second is the chargeback). ex: ["19827356981273"]
-    - created [datetime.datetime, default None]: creation datetime for the transfer. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
-    - updated [datetime.datetime, default None]: latest update datetime for the transfer. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
+    - external_id [string]: url safe string that must be unique among all your transfers. Duplicated external_ids will cause failures. By default, this parameter will block any transfer that repeats amount and receiver information on the same date. ex: "my-internal-id-123456"
+    - scheduled [datetime.date, datetime.datetime or string]: date or datetime when the transfer will be processed. May be pushed to next business day if necessary. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
+    - description [string]: optional description to override default description to be shown in the bank statement. ex: "Payment for service #1234"
+    - id [string]: unique id returned when the transfer is created. ex: "5656565656565656"
+    - fee [integer]: fee charged when the Transfer is processed. ex: 200 (= R$ 2.00)
+    - status [string]: current transfer status. ex: "success" or "failed"
+    - transaction_ids [list of strings]: ledger Transaction IDs linked to this Transfer (if there are two, the second is the chargeback). ex: ["19827356981273"]
+    - created [datetime.datetime]: creation datetime for the transfer. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
+    - updated [datetime.datetime]: latest update datetime for the transfer. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     """
 
     def __init__(self, name, tax_id, bank_code, branch_code, account_number, amount=None, account_type=None,